shell: split the commands.h for refactor (#278)
